Fix delete-related errors

+++ app/controllers/products_controller.rb
@@ -14,10 +14,15 @@ class ProductsController < ApplicationController
   end
 
   def destroy
-    if @product.destroy
-      redirect_to mypage_index_path(current_user.id), notice: '商品を削除しました'
+    if user_signed_in? && current_user.id == @product.user_id
+      if @product.destroy
+        redirect_to mypage_index_path(current_user.id), notice: '商品を削除しました'
+      else
+        render @product
+      end
     else
-      render @product
+      # 直前の画面に戻る
+      return_back and return
     end
   end
